Graphics.X11.ExtraTypes.XorgDefault

Description

This file is generated based on X.org includes. It contains the keysyms for XorgDefault. Because there are so many entries and each one is checked for #ifdef twice, the C preprocessor part takes a long time.

Documentation

xK_ISO_Lock :: KeySym

xK_ISO_Level2_Latch :: KeySym

xK_ISO_Level3_Shift :: KeySym

xK_ISO_Level3_Latch :: KeySym

xK_ISO_Level3_Lock :: KeySym

xK_ISO_Level5_Shift :: KeySym

xK_ISO_Level5_Latch :: KeySym

xK_ISO_Level5_Lock :: KeySym

xK_ISO_Group_Shift :: KeySym

xK_ISO_Group_Latch :: KeySym

xK_ISO_Group_Lock :: KeySym

xK_ISO_Next_Group :: KeySym

xK_ISO_Next_Group_Lock :: KeySym

xK_ISO_Prev_Group :: KeySym

xK_ISO_Prev_Group_Lock :: KeySym

xK_ISO_First_Group :: KeySym

xK_ISO_First_Group_Lock :: KeySym

xK_ISO_Last_Group :: KeySym

xK_ISO_Last_Group_Lock :: KeySym

xK_ISO_Left_Tab :: KeySym

xK_ISO_Move_Line_Up :: KeySym

xK_ISO_Move_Line_Down :: KeySym

xK_ISO_Partial_Line_Up :: KeySym

xK_ISO_Partial_Line_Down :: KeySym

xK_ISO_Partial_Space_Left :: KeySym

xK_ISO_Partial_Space_Right :: KeySym

xK_ISO_Set_Margin_Left :: KeySym

xK_ISO_Set_Margin_Right :: KeySym

xK_ISO_Release_Margin_Left :: KeySym

xK_ISO_Release_Margin_Right :: KeySym

xK_ISO_Release_Both_Margins :: KeySym

xK_ISO_Fast_Cursor_Left :: KeySym

xK_ISO_Fast_Cursor_Right :: KeySym

xK_ISO_Fast_Cursor_Up :: KeySym

xK_ISO_Fast_Cursor_Down :: KeySym

xK_ISO_Continuous_Underline :: KeySym

xK_ISO_Discontinuous_Underline :: KeySym

xK_ISO_Emphasize :: KeySym

xK_ISO_Center_Object :: KeySym

xK_ISO_Enter :: KeySym

xK_dead_grave :: KeySym

xK_dead_acute :: KeySym

xK_dead_circumflex :: KeySym

xK_dead_tilde :: KeySym

xK_dead_macron :: KeySym

xK_dead_breve :: KeySym

xK_dead_abovedot :: KeySym

xK_dead_diaeresis :: KeySym

xK_dead_abovering :: KeySym

xK_dead_doubleacute :: KeySym

xK_dead_caron :: KeySym

xK_dead_cedilla :: KeySym

xK_dead_ogonek :: KeySym

xK_dead_iota :: KeySym

xK_dead_voiced_sound :: KeySym

xK_dead_semivoiced_sound :: KeySym

xK_dead_belowdot :: KeySym

xK_dead_hook :: KeySym

xK_dead_horn :: KeySym

xK_dead_stroke :: KeySym

xK_dead_abovecomma :: KeySym

xK_dead_psili :: KeySym

xK_dead_abovereversedcomma :: KeySym

xK_dead_dasia :: KeySym

xK_First_Virtual_Screen :: KeySym

xK_Prev_Virtual_Screen :: KeySym

xK_Next_Virtual_Screen :: KeySym

xK_Last_Virtual_Screen :: KeySym

xK_Terminate_Server :: KeySym

xK_AccessX_Enable :: KeySym

xK_AccessX_Feedback_Enable :: KeySym

xK_RepeatKeys_Enable :: KeySym

xK_SlowKeys_Enable :: KeySym

xK_BounceKeys_Enable :: KeySym

xK_StickyKeys_Enable :: KeySym

xK_MouseKeys_Enable :: KeySym

xK_MouseKeys_Accel_Enable :: KeySym

xK_Overlay1_Enable :: KeySym

xK_Overlay2_Enable :: KeySym

xK_AudibleBell_Enable :: KeySym

xK_Pointer_Left :: KeySym

xK_Pointer_Right :: KeySym

xK_Pointer_Up :: KeySym

xK_Pointer_Down :: KeySym

xK_Pointer_UpLeft :: KeySym

xK_Pointer_UpRight :: KeySym

xK_Pointer_DownLeft :: KeySym

xK_Pointer_DownRight :: KeySym

xK_Pointer_Button_Dflt :: KeySym

xK_Pointer_Button1 :: KeySym

xK_Pointer_Button2 :: KeySym

xK_Pointer_Button3 :: KeySym

xK_Pointer_Button4 :: KeySym

xK_Pointer_Button5 :: KeySym

xK_Pointer_DblClick_Dflt :: KeySym

xK_Pointer_DblClick1 :: KeySym

xK_Pointer_DblClick2 :: KeySym

xK_Pointer_DblClick3 :: KeySym

xK_Pointer_DblClick4 :: KeySym

xK_Pointer_DblClick5 :: KeySym

xK_Pointer_Drag_Dflt :: KeySym

xK_Pointer_Drag1 :: KeySym

xK_Pointer_Drag2 :: KeySym

xK_Pointer_Drag3 :: KeySym

xK_Pointer_Drag4 :: KeySym

xK_Pointer_Drag5 :: KeySym

xK_Pointer_EnableKeys :: KeySym

xK_Pointer_Accelerate :: KeySym

xK_Pointer_DfltBtnNext :: KeySym

xK_Pointer_DfltBtnPrev :: KeySym

xK_Aogonek :: KeySym

xK_breve :: KeySym

xK_Lstroke :: KeySym

xK_Lcaron :: KeySym

xK_Sacute :: KeySym

xK_Scaron :: KeySym

xK_Scedilla :: KeySym

xK_Tcaron :: KeySym

xK_Zacute :: KeySym

xK_Zcaron :: KeySym

xK_Zabovedot :: KeySym

xK_aogonek :: KeySym

xK_ogonek :: KeySym

xK_lstroke :: KeySym

xK_lcaron :: KeySym

xK_sacute :: KeySym

xK_caron :: KeySym

xK_scaron :: KeySym

xK_scedilla :: KeySym

xK_tcaron :: KeySym

xK_zacute :: KeySym

xK_doubleacute :: KeySym

xK_zcaron :: KeySym

xK_zabovedot :: KeySym

xK_Racute :: KeySym

xK_Abreve :: KeySym

xK_Lacute :: KeySym

xK_Cacute :: KeySym

xK_Ccaron :: KeySym

xK_Eogonek :: KeySym

xK_Ecaron :: KeySym

xK_Dcaron :: KeySym

xK_Dstroke :: KeySym

xK_Nacute :: KeySym

xK_Ncaron :: KeySym

xK_Odoubleacute :: KeySym

xK_Rcaron :: KeySym

xK_Uring :: KeySym

xK_Udoubleacute :: KeySym

xK_Tcedilla :: KeySym

xK_racute :: KeySym

xK_abreve :: KeySym

xK_lacute :: KeySym

xK_cacute :: KeySym

xK_ccaron :: KeySym

xK_eogonek :: KeySym

xK_ecaron :: KeySym

xK_dcaron :: KeySym

xK_dstroke :: KeySym

xK_nacute :: KeySym

xK_ncaron :: KeySym

xK_odoubleacute :: KeySym

xK_udoubleacute :: KeySym

xK_rcaron :: KeySym

xK_uring :: KeySym

xK_tcedilla :: KeySym

xK_abovedot :: KeySym

xK_Hstroke :: KeySym

xK_Hcircumflex :: KeySym

xK_Iabovedot :: KeySym

xK_Gbreve :: KeySym

xK_Jcircumflex :: KeySym

xK_hstroke :: KeySym

xK_hcircumflex :: KeySym

xK_idotless :: KeySym

xK_gbreve :: KeySym

xK_jcircumflex :: KeySym

xK_Cabovedot :: KeySym

xK_Ccircumflex :: KeySym

xK_Gabovedot :: KeySym

xK_Gcircumflex :: KeySym

xK_Ubreve :: KeySym

xK_Scircumflex :: KeySym

xK_cabovedot :: KeySym

xK_ccircumflex :: KeySym

xK_gabovedot :: KeySym

xK_gcircumflex :: KeySym

xK_ubreve :: KeySym

xK_scircumflex :: KeySym

xK_kra :: KeySym

xK_kappa :: KeySym

xK_Rcedilla :: KeySym

xK_Itilde :: KeySym

xK_Lcedilla :: KeySym

xK_Emacron :: KeySym

xK_Gcedilla :: KeySym

xK_Tslash :: KeySym

xK_rcedilla :: KeySym

xK_itilde :: KeySym

xK_lcedilla :: KeySym

xK_emacron :: KeySym

xK_gcedilla :: KeySym

xK_tslash :: KeySym

xK_ENG :: KeySym

xK_eng :: KeySym

xK_Amacron :: KeySym

xK_Iogonek :: KeySym

xK_Eabovedot :: KeySym

xK_Imacron :: KeySym

xK_Ncedilla :: KeySym

xK_Omacron :: KeySym

xK_Kcedilla :: KeySym

xK_Uogonek :: KeySym

xK_Utilde :: KeySym

xK_Umacron :: KeySym

xK_amacron :: KeySym

xK_iogonek :: KeySym

xK_eabovedot :: KeySym

xK_imacron :: KeySym

xK_ncedilla :: KeySym

xK_omacron :: KeySym

xK_kcedilla :: KeySym

xK_uogonek :: KeySym

xK_utilde :: KeySym

xK_umacron :: KeySym

xK_Babovedot :: KeySym

xK_babovedot :: KeySym

xK_Dabovedot :: KeySym

xK_Wgrave :: KeySym

xK_Wacute :: KeySym

xK_dabovedot :: KeySym

xK_Ygrave :: KeySym

xK_Fabovedot :: KeySym

xK_fabovedot :: KeySym

xK_Mabovedot :: KeySym

xK_mabovedot :: KeySym

xK_Pabovedot :: KeySym

xK_wgrave :: KeySym

xK_pabovedot :: KeySym

xK_wacute :: KeySym

xK_Sabovedot :: KeySym

xK_ygrave :: KeySym

xK_Wdiaeresis :: KeySym

xK_wdiaeresis :: KeySym

xK_sabovedot :: KeySym

xK_Wcircumflex :: KeySym

xK_Tabovedot :: KeySym

xK_Ycircumflex :: KeySym

xK_wcircumflex :: KeySym

xK_tabovedot :: KeySym

xK_ycircumflex :: KeySym

xK_OE :: KeySym

xK_oe :: KeySym

xK_Ydiaeresis :: KeySym

xK_overline :: KeySym

xK_kana_fullstop :: KeySym

xK_kana_openingbracket :: KeySym

xK_kana_closingbracket :: KeySym

xK_kana_comma :: KeySym

xK_kana_conjunctive :: KeySym

xK_kana_middledot :: KeySym

xK_kana_WO :: KeySym

xK_kana_a :: KeySym

xK_kana_i :: KeySym

xK_kana_u :: KeySym

xK_kana_e :: KeySym

xK_kana_o :: KeySym

xK_kana_ya :: KeySym

xK_kana_yu :: KeySym

xK_kana_yo :: KeySym

xK_kana_tsu :: KeySym

xK_kana_tu :: KeySym

xK_prolongedsound :: KeySym

xK_kana_A :: KeySym

xK_kana_I :: KeySym

xK_kana_U :: KeySym

xK_kana_E :: KeySym

xK_kana_O :: KeySym

xK_kana_KA :: KeySym

xK_kana_KI :: KeySym

xK_kana_KU :: KeySym

xK_kana_KE :: KeySym

xK_kana_KO :: KeySym

xK_kana_SA :: KeySym

xK_kana_SHI :: KeySym

xK_kana_SU :: KeySym

xK_kana_SE :: KeySym

xK_kana_SO :: KeySym

xK_kana_TA :: KeySym

xK_kana_CHI :: KeySym

xK_kana_TI :: KeySym

xK_kana_TSU :: KeySym

xK_kana_TU :: KeySym

xK_kana_TE :: KeySym

xK_kana_TO :: KeySym

xK_kana_NA :: KeySym

xK_kana_NI :: KeySym

xK_kana_NU :: KeySym

xK_kana_NE :: KeySym

xK_kana_NO :: KeySym

xK_kana_HA :: KeySym

xK_kana_HI :: KeySym

xK_kana_FU :: KeySym

xK_kana_HU :: KeySym

xK_kana_HE :: KeySym

xK_kana_HO :: KeySym

xK_kana_MA :: KeySym

xK_kana_MI :: KeySym

xK_kana_MU :: KeySym

xK_kana_ME :: KeySym

xK_kana_MO :: KeySym

xK_kana_YA :: KeySym

xK_kana_YU :: KeySym

xK_kana_YO :: KeySym

xK_kana_RA :: KeySym

xK_kana_RI :: KeySym

xK_kana_RU :: KeySym

xK_kana_RE :: KeySym

xK_kana_RO :: KeySym

xK_kana_WA :: KeySym

xK_kana_N :: KeySym

xK_voicedsound :: KeySym

xK_semivoicedsound :: KeySym

xK_kana_switch :: KeySym

xK_Farsi_0 :: KeySym

xK_Farsi_1 :: KeySym

xK_Farsi_2 :: KeySym

xK_Farsi_3 :: KeySym

xK_Farsi_4 :: KeySym

xK_Farsi_5 :: KeySym

xK_Farsi_6 :: KeySym

xK_Farsi_7 :: KeySym

xK_Farsi_8 :: KeySym

xK_Farsi_9 :: KeySym

xK_Arabic_percent :: KeySym

xK_Arabic_superscript_alef :: KeySym

xK_Arabic_tteh :: KeySym

xK_Arabic_peh :: KeySym

xK_Arabic_tcheh :: KeySym

xK_Arabic_ddal :: KeySym

xK_Arabic_rreh :: KeySym

xK_Arabic_comma :: KeySym

xK_Arabic_fullstop :: KeySym

xK_Arabic_0 :: KeySym

xK_Arabic_1 :: KeySym

xK_Arabic_2 :: KeySym

xK_Arabic_3 :: KeySym

xK_Arabic_4 :: KeySym

xK_Arabic_5 :: KeySym

xK_Arabic_6 :: KeySym

xK_Arabic_7 :: KeySym

xK_Arabic_8 :: KeySym

xK_Arabic_9 :: KeySym

xK_Arabic_semicolon :: KeySym

xK_Arabic_question_mark :: KeySym

xK_Arabic_hamza :: KeySym

xK_Arabic_maddaonalef :: KeySym

xK_Arabic_hamzaonalef :: KeySym

xK_Arabic_hamzaonwaw :: KeySym

xK_Arabic_hamzaunderalef :: KeySym

xK_Arabic_hamzaonyeh :: KeySym

xK_Arabic_alef :: KeySym

xK_Arabic_beh :: KeySym

xK_Arabic_tehmarbuta :: KeySym

xK_Arabic_teh :: KeySym

xK_Arabic_theh :: KeySym

xK_Arabic_jeem :: KeySym

xK_Arabic_hah :: KeySym

xK_Arabic_khah :: KeySym

xK_Arabic_dal :: KeySym

xK_Arabic_thal :: KeySym

xK_Arabic_ra :: KeySym

xK_Arabic_zain :: KeySym

xK_Arabic_seen :: KeySym

xK_Arabic_sheen :: KeySym

xK_Arabic_sad :: KeySym

xK_Arabic_dad :: KeySym

xK_Arabic_tah :: KeySym

xK_Arabic_zah :: KeySym

xK_Arabic_ain :: KeySym

xK_Arabic_ghain :: KeySym

xK_Arabic_tatweel :: KeySym

xK_Arabic_feh :: KeySym

xK_Arabic_qaf :: KeySym

xK_Arabic_kaf :: KeySym

xK_Arabic_lam :: KeySym

xK_Arabic_meem :: KeySym

xK_Arabic_noon :: KeySym

xK_Arabic_ha :: KeySym

xK_Arabic_heh :: KeySym

xK_Arabic_waw :: KeySym

xK_Arabic_alefmaksura :: KeySym

xK_Arabic_yeh :: KeySym

xK_Arabic_fathatan :: KeySym

xK_Arabic_dammatan :: KeySym

xK_Arabic_kasratan :: KeySym

xK_Arabic_fatha :: KeySym

xK_Arabic_damma :: KeySym

xK_Arabic_kasra :: KeySym

xK_Arabic_shadda :: KeySym

xK_Arabic_sukun :: KeySym

xK_Arabic_madda_above :: KeySym

xK_Arabic_hamza_above :: KeySym

xK_Arabic_hamza_below :: KeySym

xK_Arabic_jeh :: KeySym

xK_Arabic_veh :: KeySym

xK_Arabic_keheh :: KeySym

xK_Arabic_gaf :: KeySym

xK_Arabic_noon_ghunna :: KeySym

xK_Arabic_heh_doachashmee :: KeySym

xK_Farsi_yeh :: KeySym

xK_Arabic_farsi_yeh :: KeySym

xK_Arabic_yeh_baree :: KeySym

xK_Arabic_heh_goal :: KeySym

xK_Arabic_switch :: KeySym

xK_Cyrillic_GHE_bar :: KeySym

xK_Cyrillic_ghe_bar :: KeySym

xK_Cyrillic_ZHE_descender :: KeySym

xK_Cyrillic_zhe_descender :: KeySym

xK_Cyrillic_KA_descender :: KeySym

xK_Cyrillic_ka_descender :: KeySym

xK_Cyrillic_KA_vertstroke :: KeySym

xK_Cyrillic_ka_vertstroke :: KeySym

xK_Cyrillic_EN_descender :: KeySym

xK_Cyrillic_en_descender :: KeySym

xK_Cyrillic_U_straight :: KeySym

xK_Cyrillic_u_straight :: KeySym

xK_Cyrillic_U_straight_bar :: KeySym

xK_Cyrillic_u_straight_bar :: KeySym

xK_Cyrillic_HA_descender :: KeySym

xK_Cyrillic_ha_descender :: KeySym

xK_Cyrillic_CHE_descender :: KeySym

xK_Cyrillic_che_descender :: KeySym

xK_Cyrillic_CHE_vertstroke :: KeySym

xK_Cyrillic_che_vertstroke :: KeySym

xK_Cyrillic_SHHA :: KeySym

xK_Cyrillic_shha :: KeySym

xK_Cyrillic_SCHWA :: KeySym

xK_Cyrillic_schwa :: KeySym

xK_Cyrillic_I_macron :: KeySym

xK_Cyrillic_i_macron :: KeySym

xK_Cyrillic_O_bar :: KeySym

xK_Cyrillic_o_bar :: KeySym

xK_Cyrillic_U_macron :: KeySym

xK_Cyrillic_u_macron :: KeySym

xK_Serbian_dje :: KeySym

xK_Macedonia_gje :: KeySym

xK_Cyrillic_io :: KeySym

xK_Ukrainian_ie :: KeySym

xK_Ukranian_je :: KeySym

xK_Macedonia_dse :: KeySym

xK_Ukrainian_i :: KeySym

xK_Ukranian_i :: KeySym

xK_Ukrainian_yi :: KeySym

xK_Ukranian_yi :: KeySym

xK_Cyrillic_je :: KeySym

xK_Serbian_je :: KeySym

xK_Cyrillic_lje :: KeySym

xK_Serbian_lje :: KeySym

xK_Cyrillic_nje :: KeySym

xK_Serbian_nje :: KeySym

xK_Serbian_tshe :: KeySym

xK_Macedonia_kje :: KeySym

xK_Ukrainian_ghe_with_upturn :: KeySym

xK_Byelorussian_shortu :: KeySym

xK_Cyrillic_dzhe :: KeySym

xK_Serbian_dze :: KeySym

xK_numerosign :: KeySym

xK_Serbian_DJE :: KeySym

xK_Macedonia_GJE :: KeySym

xK_Cyrillic_IO :: KeySym

xK_Ukrainian_IE :: KeySym

xK_Ukranian_JE :: KeySym

xK_Macedonia_DSE :: KeySym

xK_Ukrainian_I :: KeySym

xK_Ukranian_I :: KeySym

xK_Ukrainian_YI :: KeySym

xK_Ukranian_YI :: KeySym

xK_Cyrillic_JE :: KeySym

xK_Serbian_JE :: KeySym

xK_Cyrillic_LJE :: KeySym

xK_Serbian_LJE :: KeySym

xK_Cyrillic_NJE :: KeySym

xK_Serbian_NJE :: KeySym

xK_Serbian_TSHE :: KeySym

xK_Macedonia_KJE :: KeySym

xK_Ukrainian_GHE_WITH_UPTURN :: KeySym

xK_Byelorussian_SHORTU :: KeySym

xK_Cyrillic_DZHE :: KeySym

xK_Serbian_DZE :: KeySym

xK_Cyrillic_yu :: KeySym

xK_Cyrillic_a :: KeySym

xK_Cyrillic_be :: KeySym

xK_Cyrillic_tse :: KeySym

xK_Cyrillic_de :: KeySym

xK_Cyrillic_ie :: KeySym

xK_Cyrillic_ef :: KeySym

xK_Cyrillic_ghe :: KeySym

xK_Cyrillic_ha :: KeySym

xK_Cyrillic_i :: KeySym

xK_Cyrillic_shorti :: KeySym

xK_Cyrillic_ka :: KeySym

xK_Cyrillic_el :: KeySym

xK_Cyrillic_em :: KeySym

xK_Cyrillic_en :: KeySym

xK_Cyrillic_o :: KeySym

xK_Cyrillic_pe :: KeySym

xK_Cyrillic_ya :: KeySym

xK_Cyrillic_er :: KeySym

xK_Cyrillic_es :: KeySym

xK_Cyrillic_te :: KeySym

xK_Cyrillic_u :: KeySym

xK_Cyrillic_zhe :: KeySym

xK_Cyrillic_ve :: KeySym

xK_Cyrillic_softsign :: KeySym

xK_Cyrillic_yeru :: KeySym

xK_Cyrillic_ze :: KeySym

xK_Cyrillic_sha :: KeySym

xK_Cyrillic_e :: KeySym

xK_Cyrillic_shcha :: KeySym

xK_Cyrillic_che :: KeySym

xK_Cyrillic_hardsign :: KeySym

xK_Cyrillic_YU :: KeySym

xK_Cyrillic_A :: KeySym

xK_Cyrillic_BE :: KeySym

xK_Cyrillic_TSE :: KeySym

xK_Cyrillic_DE :: KeySym

xK_Cyrillic_IE :: KeySym

xK_Cyrillic_EF :: KeySym

xK_Cyrillic_GHE :: KeySym

xK_Cyrillic_HA :: KeySym

xK_Cyrillic_I :: KeySym

xK_Cyrillic_SHORTI :: KeySym

xK_Cyrillic_KA :: KeySym

xK_Cyrillic_EL :: KeySym

xK_Cyrillic_EM :: KeySym

xK_Cyrillic_EN :: KeySym

xK_Cyrillic_O :: KeySym

xK_Cyrillic_PE :: KeySym

xK_Cyrillic_YA :: KeySym

xK_Cyrillic_ER :: KeySym

xK_Cyrillic_ES :: KeySym

xK_Cyrillic_TE :: KeySym

xK_Cyrillic_U :: KeySym

xK_Cyrillic_ZHE :: KeySym

xK_Cyrillic_VE :: KeySym

xK_Cyrillic_SOFTSIGN :: KeySym

xK_Cyrillic_YERU :: KeySym

xK_Cyrillic_ZE :: KeySym

xK_Cyrillic_SHA :: KeySym

xK_Cyrillic_E :: KeySym

xK_Cyrillic_SHCHA :: KeySym

xK_Cyrillic_CHE :: KeySym

xK_Cyrillic_HARDSIGN :: KeySym

xK_Greek_ALPHAaccent :: KeySym

xK_Greek_EPSILONaccent :: KeySym

xK_Greek_ETAaccent :: KeySym

xK_Greek_IOTAaccent :: KeySym

xK_Greek_IOTAdieresis :: KeySym

xK_Greek_IOTAdiaeresis :: KeySym

xK_Greek_OMICRONaccent :: KeySym

xK_Greek_UPSILONaccent :: KeySym

xK_Greek_UPSILONdieresis :: KeySym

xK_Greek_OMEGAaccent :: KeySym

xK_Greek_accentdieresis :: KeySym

xK_Greek_horizbar :: KeySym

xK_Greek_alphaaccent :: KeySym

xK_Greek_epsilonaccent :: KeySym

xK_Greek_etaaccent :: KeySym

xK_Greek_iotaaccent :: KeySym

xK_Greek_iotadieresis :: KeySym

xK_Greek_iotaaccentdieresis :: KeySym

xK_Greek_omicronaccent :: KeySym

xK_Greek_upsilonaccent :: KeySym

xK_Greek_upsilondieresis :: KeySym

xK_Greek_upsilonaccentdieresis :: KeySym

xK_Greek_omegaaccent :: KeySym

xK_Greek_ALPHA :: KeySym

xK_Greek_BETA :: KeySym

xK_Greek_GAMMA :: KeySym

xK_Greek_DELTA :: KeySym

xK_Greek_EPSILON :: KeySym

xK_Greek_ZETA :: KeySym

xK_Greek_ETA :: KeySym

xK_Greek_THETA :: KeySym

xK_Greek_IOTA :: KeySym

xK_Greek_KAPPA :: KeySym

xK_Greek_LAMDA :: KeySym

xK_Greek_LAMBDA :: KeySym

xK_Greek_MU :: KeySym

xK_Greek_NU :: KeySym

xK_Greek_XI :: KeySym

xK_Greek_OMICRON :: KeySym

xK_Greek_PI :: KeySym

xK_Greek_RHO :: KeySym

xK_Greek_SIGMA :: KeySym

xK_Greek_TAU :: KeySym

xK_Greek_UPSILON :: KeySym

xK_Greek_PHI :: KeySym

xK_Greek_CHI :: KeySym

xK_Greek_PSI :: KeySym

xK_Greek_OMEGA :: KeySym

xK_Greek_alpha :: KeySym

xK_Greek_beta :: KeySym

xK_Greek_gamma :: KeySym

xK_Greek_delta :: KeySym

xK_Greek_epsilon :: KeySym

xK_Greek_zeta :: KeySym

xK_Greek_eta :: KeySym

xK_Greek_theta :: KeySym

xK_Greek_iota :: KeySym

xK_Greek_kappa :: KeySym

xK_Greek_lamda :: KeySym

xK_Greek_lambda :: KeySym

xK_Greek_mu :: KeySym

xK_Greek_nu :: KeySym

xK_Greek_xi :: KeySym

xK_Greek_omicron :: KeySym

xK_Greek_pi :: KeySym

xK_Greek_rho :: KeySym

xK_Greek_sigma :: KeySym

xK_Greek_finalsmallsigma :: KeySym

xK_Greek_tau :: KeySym

xK_Greek_upsilon :: KeySym

xK_Greek_phi :: KeySym

xK_Greek_chi :: KeySym

xK_Greek_psi :: KeySym

xK_Greek_omega :: KeySym

xK_Greek_switch :: KeySym

xK_hebrew_doublelowline :: KeySym

xK_hebrew_aleph :: KeySym

xK_hebrew_bet :: KeySym

xK_hebrew_beth :: KeySym

xK_hebrew_gimel :: KeySym

xK_hebrew_gimmel :: KeySym

xK_hebrew_dalet :: KeySym

xK_hebrew_daleth :: KeySym

xK_hebrew_he :: KeySym

xK_hebrew_waw :: KeySym

xK_hebrew_zain :: KeySym

xK_hebrew_zayin :: KeySym

xK_hebrew_chet :: KeySym

xK_hebrew_het :: KeySym

xK_hebrew_tet :: KeySym

xK_hebrew_teth :: KeySym

xK_hebrew_yod :: KeySym

xK_hebrew_finalkaph :: KeySym

xK_hebrew_kaph :: KeySym

xK_hebrew_lamed :: KeySym

xK_hebrew_finalmem :: KeySym

xK_hebrew_mem :: KeySym

xK_hebrew_finalnun :: KeySym

xK_hebrew_nun :: KeySym

xK_hebrew_samech :: KeySym

xK_hebrew_samekh :: KeySym

xK_hebrew_ayin :: KeySym

xK_hebrew_finalpe :: KeySym

xK_hebrew_pe :: KeySym

xK_hebrew_finalzade :: KeySym

xK_hebrew_finalzadi :: KeySym

xK_hebrew_zade :: KeySym

xK_hebrew_zadi :: KeySym

xK_hebrew_qoph :: KeySym

xK_hebrew_kuf :: KeySym

xK_hebrew_resh :: KeySym

xK_hebrew_shin :: KeySym

xK_hebrew_taw :: KeySym

xK_hebrew_taf :: KeySym

xK_Hebrew_switch :: KeySym

xK_Thai_kokai :: KeySym

xK_Thai_khokhai :: KeySym

xK_Thai_khokhuat :: KeySym

xK_Thai_khokhwai :: KeySym

xK_Thai_khokhon :: KeySym

xK_Thai_khorakhang :: KeySym

xK_Thai_ngongu :: KeySym

xK_Thai_chochan :: KeySym

xK_Thai_choching :: KeySym

xK_Thai_chochang :: KeySym

xK_Thai_soso :: KeySym

xK_Thai_chochoe :: KeySym

xK_Thai_yoying :: KeySym

xK_Thai_dochada :: KeySym

xK_Thai_topatak :: KeySym

xK_Thai_thothan :: KeySym

xK_Thai_thonangmontho :: KeySym

xK_Thai_thophuthao :: KeySym

xK_Thai_nonen :: KeySym

xK_Thai_dodek :: KeySym

xK_Thai_totao :: KeySym

xK_Thai_thothung :: KeySym

xK_Thai_thothahan :: KeySym

xK_Thai_thothong :: KeySym

xK_Thai_nonu :: KeySym

xK_Thai_bobaimai :: KeySym

xK_Thai_popla :: KeySym

xK_Thai_phophung :: KeySym

xK_Thai_fofa :: KeySym

xK_Thai_phophan :: KeySym

xK_Thai_fofan :: KeySym

xK_Thai_phosamphao :: KeySym

xK_Thai_moma :: KeySym

xK_Thai_yoyak :: KeySym

xK_Thai_rorua :: KeySym

xK_Thai_ru :: KeySym

xK_Thai_loling :: KeySym

xK_Thai_lu :: KeySym

xK_Thai_wowaen :: KeySym

xK_Thai_sosala :: KeySym

xK_Thai_sorusi :: KeySym

xK_Thai_sosua :: KeySym

xK_Thai_hohip :: KeySym

xK_Thai_lochula :: KeySym

xK_Thai_oang :: KeySym

xK_Thai_honokhuk :: KeySym

xK_Thai_paiyannoi :: KeySym

xK_Thai_saraa :: KeySym

xK_Thai_maihanakat :: KeySym

xK_Thai_saraaa :: KeySym

xK_Thai_saraam :: KeySym

xK_Thai_sarai :: KeySym

xK_Thai_saraii :: KeySym

xK_Thai_saraue :: KeySym

xK_Thai_sarauee :: KeySym

xK_Thai_sarau :: KeySym

xK_Thai_sarauu :: KeySym

xK_Thai_phinthu :: KeySym

xK_Thai_maihanakat_maitho :: KeySym

xK_Thai_baht :: KeySym

xK_Thai_sarae :: KeySym

xK_Thai_saraae :: KeySym

xK_Thai_sarao :: KeySym

xK_Thai_saraaimaimuan :: KeySym

xK_Thai_saraaimaimalai :: KeySym

xK_Thai_lakkhangyao :: KeySym

xK_Thai_maiyamok :: KeySym

xK_Thai_maitaikhu :: KeySym

xK_Thai_maiek :: KeySym

xK_Thai_maitho :: KeySym

xK_Thai_maitri :: KeySym

xK_Thai_maichattawa :: KeySym

xK_Thai_thanthakhat :: KeySym

xK_Thai_nikhahit :: KeySym

xK_Thai_leksun :: KeySym

xK_Thai_leknung :: KeySym

xK_Thai_leksong :: KeySym

xK_Thai_leksam :: KeySym

xK_Thai_leksi :: KeySym

xK_Thai_lekha :: KeySym

xK_Thai_lekhok :: KeySym

xK_Thai_lekchet :: KeySym

xK_Thai_lekpaet :: KeySym

xK_Thai_lekkao :: KeySym

xK_Hangul :: KeySym

xK_Hangul_Start :: KeySym

xK_Hangul_End :: KeySym

xK_Hangul_Hanja :: KeySym

xK_Hangul_Jamo :: KeySym

xK_Hangul_Romaja :: KeySym

xK_Hangul_Codeinput :: KeySym

xK_Hangul_Jeonja :: KeySym

xK_Hangul_Banja :: KeySym

xK_Hangul_PreHanja :: KeySym

xK_Hangul_PostHanja :: KeySym

xK_Hangul_SingleCandidate :: KeySym

xK_Hangul_MultipleCandidate :: KeySym

xK_Hangul_PreviousCandidate :: KeySym

xK_Hangul_Special :: KeySym

xK_Hangul_switch :: KeySym

xK_Hangul_Kiyeog :: KeySym

xK_Hangul_SsangKiyeog :: KeySym

xK_Hangul_KiyeogSios :: KeySym

xK_Hangul_Nieun :: KeySym

xK_Hangul_NieunJieuj :: KeySym

xK_Hangul_NieunHieuh :: KeySym

xK_Hangul_Dikeud :: KeySym

xK_Hangul_SsangDikeud :: KeySym

xK_Hangul_Rieul :: KeySym

xK_Hangul_RieulKiyeog :: KeySym

xK_Hangul_RieulMieum :: KeySym

xK_Hangul_RieulPieub :: KeySym

xK_Hangul_RieulSios :: KeySym

xK_Hangul_RieulTieut :: KeySym

xK_Hangul_RieulPhieuf :: KeySym

xK_Hangul_RieulHieuh :: KeySym

xK_Hangul_Mieum :: KeySym

xK_Hangul_Pieub :: KeySym

xK_Hangul_SsangPieub :: KeySym

xK_Hangul_PieubSios :: KeySym

xK_Hangul_Sios :: KeySym

xK_Hangul_SsangSios :: KeySym

xK_Hangul_Ieung :: KeySym

xK_Hangul_Jieuj :: KeySym

xK_Hangul_SsangJieuj :: KeySym

xK_Hangul_Cieuc :: KeySym

xK_Hangul_Khieuq :: KeySym

xK_Hangul_Tieut :: KeySym

xK_Hangul_Phieuf :: KeySym

xK_Hangul_Hieuh :: KeySym

xK_Hangul_A :: KeySym

xK_Hangul_AE :: KeySym

xK_Hangul_YA :: KeySym

xK_Hangul_YAE :: KeySym

xK_Hangul_EO :: KeySym

xK_Hangul_E :: KeySym

xK_Hangul_YEO :: KeySym

xK_Hangul_YE :: KeySym

xK_Hangul_O :: KeySym

xK_Hangul_WA :: KeySym

xK_Hangul_WAE :: KeySym

xK_Hangul_OE :: KeySym

xK_Hangul_YO :: KeySym

xK_Hangul_U :: KeySym

xK_Hangul_WEO :: KeySym

xK_Hangul_WE :: KeySym

xK_Hangul_WI :: KeySym

xK_Hangul_YU :: KeySym

xK_Hangul_EU :: KeySym

xK_Hangul_YI :: KeySym

xK_Hangul_I :: KeySym

xK_Hangul_J_Kiyeog :: KeySym

xK_Hangul_J_SsangKiyeog :: KeySym

xK_Hangul_J_KiyeogSios :: KeySym

xK_Hangul_J_Nieun :: KeySym

xK_Hangul_J_NieunJieuj :: KeySym

xK_Hangul_J_NieunHieuh :: KeySym

xK_Hangul_J_Dikeud :: KeySym

xK_Hangul_J_Rieul :: KeySym

xK_Hangul_J_RieulKiyeog :: KeySym

xK_Hangul_J_RieulMieum :: KeySym

xK_Hangul_J_RieulPieub :: KeySym

xK_Hangul_J_RieulSios :: KeySym

xK_Hangul_J_RieulTieut :: KeySym

xK_Hangul_J_RieulPhieuf :: KeySym

xK_Hangul_J_RieulHieuh :: KeySym

xK_Hangul_J_Mieum :: KeySym

xK_Hangul_J_Pieub :: KeySym

xK_Hangul_J_PieubSios :: KeySym

xK_Hangul_J_Sios :: KeySym

xK_Hangul_J_SsangSios :: KeySym

xK_Hangul_J_Ieung :: KeySym

xK_Hangul_J_Jieuj :: KeySym

xK_Hangul_J_Cieuc :: KeySym

xK_Hangul_J_Khieuq :: KeySym

xK_Hangul_J_Tieut :: KeySym

xK_Hangul_J_Phieuf :: KeySym

xK_Hangul_J_Hieuh :: KeySym

xK_Hangul_RieulYeorinHieuh :: KeySym

xK_Hangul_SunkyeongeumMieum :: KeySym

xK_Hangul_SunkyeongeumPieub :: KeySym

xK_Hangul_PanSios :: KeySym

xK_Hangul_KkogjiDalrinIeung :: KeySym

xK_Hangul_SunkyeongeumPhieuf :: KeySym

xK_Hangul_YeorinHieuh :: KeySym

xK_Hangul_AraeA :: KeySym

xK_Hangul_AraeAE :: KeySym

xK_Hangul_J_PanSios :: KeySym

xK_Hangul_J_KkogjiDalrinIeung :: KeySym

xK_Hangul_J_YeorinHieuh :: KeySym

xK_Korean_Won :: KeySym

xK_Armenian_ligature_ew :: KeySym

xK_Armenian_full_stop :: KeySym

xK_Armenian_verjaket :: KeySym

xK_Armenian_separation_mark :: KeySym

xK_Armenian_but :: KeySym

xK_Armenian_hyphen :: KeySym

xK_Armenian_yentamna :: KeySym

xK_Armenian_exclam :: KeySym

xK_Armenian_amanak :: KeySym

xK_Armenian_accent :: KeySym

xK_Armenian_shesht :: KeySym

xK_Armenian_question :: KeySym

xK_Armenian_paruyk :: KeySym

xK_Armenian_AYB :: KeySym

xK_Armenian_ayb :: KeySym

xK_Armenian_BEN :: KeySym

xK_Armenian_ben :: KeySym

xK_Armenian_GIM :: KeySym

xK_Armenian_gim :: KeySym

xK_Armenian_DA :: KeySym

xK_Armenian_da :: KeySym

xK_Armenian_YECH :: KeySym

xK_Armenian_yech :: KeySym

xK_Armenian_ZA :: KeySym

xK_Armenian_za :: KeySym

xK_Armenian_E :: KeySym

xK_Armenian_e :: KeySym

xK_Armenian_AT :: KeySym

xK_Armenian_at :: KeySym

xK_Armenian_TO :: KeySym

xK_Armenian_to :: KeySym

xK_Armenian_ZHE :: KeySym

xK_Armenian_zhe :: KeySym

xK_Armenian_INI :: KeySym

xK_Armenian_ini :: KeySym

xK_Armenian_LYUN :: KeySym

xK_Armenian_lyun :: KeySym

xK_Armenian_KHE :: KeySym

xK_Armenian_khe :: KeySym

xK_Armenian_TSA :: KeySym

xK_Armenian_tsa :: KeySym

xK_Armenian_KEN :: KeySym

xK_Armenian_ken :: KeySym

xK_Armenian_HO :: KeySym

xK_Armenian_ho :: KeySym

xK_Armenian_DZA :: KeySym

xK_Armenian_dza :: KeySym

xK_Armenian_GHAT :: KeySym

xK_Armenian_ghat :: KeySym

xK_Armenian_TCHE :: KeySym

xK_Armenian_tche :: KeySym

xK_Armenian_MEN :: KeySym

xK_Armenian_men :: KeySym

xK_Armenian_HI :: KeySym

xK_Armenian_hi :: KeySym

xK_Armenian_NU :: KeySym

xK_Armenian_nu :: KeySym

xK_Armenian_SHA :: KeySym

xK_Armenian_sha :: KeySym

xK_Armenian_VO :: KeySym

xK_Armenian_vo :: KeySym

xK_Armenian_CHA :: KeySym

xK_Armenian_cha :: KeySym

xK_Armenian_PE :: KeySym

xK_Armenian_pe :: KeySym

xK_Armenian_JE :: KeySym

xK_Armenian_je :: KeySym

xK_Armenian_RA :: KeySym

xK_Armenian_ra :: KeySym

xK_Armenian_SE :: KeySym

xK_Armenian_se :: KeySym

xK_Armenian_VEV :: KeySym

xK_Armenian_vev :: KeySym

xK_Armenian_TYUN :: KeySym

xK_Armenian_tyun :: KeySym

xK_Armenian_RE :: KeySym

xK_Armenian_re :: KeySym

xK_Armenian_TSO :: KeySym

xK_Armenian_tso :: KeySym

xK_Armenian_VYUN :: KeySym

xK_Armenian_vyun :: KeySym

xK_Armenian_PYUR :: KeySym

xK_Armenian_pyur :: KeySym

xK_Armenian_KE :: KeySym

xK_Armenian_ke :: KeySym

xK_Armenian_O :: KeySym

xK_Armenian_o :: KeySym

xK_Armenian_FE :: KeySym

xK_Armenian_fe :: KeySym

xK_Armenian_apostrophe :: KeySym

xK_Georgian_an :: KeySym

xK_Georgian_ban :: KeySym

xK_Georgian_gan :: KeySym

xK_Georgian_don :: KeySym

xK_Georgian_en :: KeySym

xK_Georgian_vin :: KeySym

xK_Georgian_zen :: KeySym

xK_Georgian_tan :: KeySym

xK_Georgian_in :: KeySym

xK_Georgian_kan :: KeySym

xK_Georgian_las :: KeySym

xK_Georgian_man :: KeySym

xK_Georgian_nar :: KeySym

xK_Georgian_on :: KeySym

xK_Georgian_par :: KeySym

xK_Georgian_zhar :: KeySym

xK_Georgian_rae :: KeySym

xK_Georgian_san :: KeySym

xK_Georgian_tar :: KeySym

xK_Georgian_un :: KeySym

xK_Georgian_phar :: KeySym

xK_Georgian_khar :: KeySym

xK_Georgian_ghan :: KeySym

xK_Georgian_qar :: KeySym

xK_Georgian_shin :: KeySym

xK_Georgian_chin :: KeySym

xK_Georgian_can :: KeySym

xK_Georgian_jil :: KeySym

xK_Georgian_cil :: KeySym

xK_Georgian_char :: KeySym

xK_Georgian_xan :: KeySym

xK_Georgian_jhan :: KeySym

xK_Georgian_hae :: KeySym

xK_Georgian_he :: KeySym

xK_Georgian_hie :: KeySym

xK_Georgian_we :: KeySym

xK_Georgian_har :: KeySym

xK_Georgian_hoe :: KeySym

xK_Georgian_fi :: KeySym

xK_Xabovedot :: KeySym

xK_Ibreve :: KeySym

xK_Zstroke :: KeySym

xK_Gcaron :: KeySym

xK_Ocaron :: KeySym

xK_Obarred :: KeySym

xK_xabovedot :: KeySym

xK_ibreve :: KeySym

xK_zstroke :: KeySym

xK_gcaron :: KeySym

xK_ocaron :: KeySym

xK_obarred :: KeySym

xK_SCHWA :: KeySym

xK_schwa :: KeySym

xK_Lbelowdot :: KeySym

xK_lbelowdot :: KeySym

xK_Abelowdot :: KeySym

xK_abelowdot :: KeySym

xK_Ahook :: KeySym

xK_ahook :: KeySym

xK_Acircumflexacute :: KeySym

xK_acircumflexacute :: KeySym

xK_Acircumflexgrave :: KeySym

xK_acircumflexgrave :: KeySym

xK_Acircumflexhook :: KeySym

xK_acircumflexhook :: KeySym

xK_Acircumflextilde :: KeySym

xK_acircumflextilde :: KeySym

xK_Acircumflexbelowdot :: KeySym

xK_acircumflexbelowdot :: KeySym

xK_Abreveacute :: KeySym

xK_abreveacute :: KeySym

xK_Abrevegrave :: KeySym

xK_abrevegrave :: KeySym

xK_Abrevehook :: KeySym

xK_abrevehook :: KeySym

xK_Abrevetilde :: KeySym

xK_abrevetilde :: KeySym

xK_Abrevebelowdot :: KeySym

xK_abrevebelowdot :: KeySym

xK_Ebelowdot :: KeySym

xK_ebelowdot :: KeySym

xK_Ehook :: KeySym

xK_ehook :: KeySym

xK_Etilde :: KeySym

xK_etilde :: KeySym

xK_Ecircumflexacute :: KeySym

xK_ecircumflexacute :: KeySym

xK_Ecircumflexgrave :: KeySym

xK_ecircumflexgrave :: KeySym

xK_Ecircumflexhook :: KeySym

xK_ecircumflexhook :: KeySym

xK_Ecircumflextilde :: KeySym

xK_ecircumflextilde :: KeySym

xK_Ecircumflexbelowdot :: KeySym

xK_ecircumflexbelowdot :: KeySym

xK_Ihook :: KeySym

xK_ihook :: KeySym

xK_Ibelowdot :: KeySym

xK_ibelowdot :: KeySym

xK_Obelowdot :: KeySym

xK_obelowdot :: KeySym

xK_Ohook :: KeySym

xK_ohook :: KeySym

xK_Ocircumflexacute :: KeySym

xK_ocircumflexacute :: KeySym

xK_Ocircumflexgrave :: KeySym

xK_ocircumflexgrave :: KeySym

xK_Ocircumflexhook :: KeySym

xK_ocircumflexhook :: KeySym

xK_Ocircumflextilde :: KeySym

xK_ocircumflextilde :: KeySym

xK_Ocircumflexbelowdot :: KeySym

xK_ocircumflexbelowdot :: KeySym

xK_Ohornacute :: KeySym

xK_ohornacute :: KeySym

xK_Ohorngrave :: KeySym

xK_ohorngrave :: KeySym

xK_Ohornhook :: KeySym

xK_ohornhook :: KeySym

xK_Ohorntilde :: KeySym

xK_ohorntilde :: KeySym

xK_Ohornbelowdot :: KeySym

xK_ohornbelowdot :: KeySym

xK_Ubelowdot :: KeySym

xK_ubelowdot :: KeySym

xK_Uhook :: KeySym

xK_uhook :: KeySym

xK_Uhornacute :: KeySym

xK_uhornacute :: KeySym

xK_Uhorngrave :: KeySym

xK_uhorngrave :: KeySym

xK_Uhornhook :: KeySym

xK_uhornhook :: KeySym

xK_Uhorntilde :: KeySym

xK_uhorntilde :: KeySym

xK_Uhornbelowdot :: KeySym

xK_uhornbelowdot :: KeySym

xK_Ybelowdot :: KeySym

xK_ybelowdot :: KeySym

xK_Yhook :: KeySym

xK_yhook :: KeySym

xK_Ytilde :: KeySym

xK_ytilde :: KeySym

xK_Ohorn :: KeySym

xK_ohorn :: KeySym

xK_Uhorn :: KeySym

xK_uhorn :: KeySym

xK_EcuSign :: KeySym

xK_ColonSign :: KeySym

xK_CruzeiroSign :: KeySym

xK_FFrancSign :: KeySym

xK_LiraSign :: KeySym

xK_MillSign :: KeySym

xK_NairaSign :: KeySym

xK_PesetaSign :: KeySym

xK_RupeeSign :: KeySym

xK_WonSign :: KeySym

xK_NewSheqelSign :: KeySym

xK_DongSign :: KeySym

xK_EuroSign :: KeySym

xK_zerosuperior :: KeySym

xK_foursuperior :: KeySym

xK_fivesuperior :: KeySym

xK_sixsuperior :: KeySym

xK_sevensuperior :: KeySym

xK_eightsuperior :: KeySym

xK_ninesuperior :: KeySym

xK_zerosubscript :: KeySym

xK_onesubscript :: KeySym

xK_twosubscript :: KeySym

xK_threesubscript :: KeySym

xK_foursubscript :: KeySym

xK_fivesubscript :: KeySym

xK_sixsubscript :: KeySym

xK_sevensubscript :: KeySym

xK_eightsubscript :: KeySym

xK_ninesubscript :: KeySym

xK_partdifferential :: KeySym

xK_emptyset :: KeySym

xK_elementof :: KeySym

xK_notelementof :: KeySym

xK_containsas :: KeySym

xK_squareroot :: KeySym

xK_cuberoot :: KeySym

xK_fourthroot :: KeySym

xK_dintegral :: KeySym

xK_tintegral :: KeySym

xK_because :: KeySym

xK_approxeq :: KeySym

xK_notapproxeq :: KeySym

xK_notidentical :: KeySym

xK_stricteq :: KeySym

xK_braille_dot_1 :: KeySym

xK_braille_dot_2 :: KeySym

xK_braille_dot_3 :: KeySym

xK_braille_dot_4 :: KeySym

xK_braille_dot_5 :: KeySym

xK_braille_dot_6 :: KeySym

xK_braille_dot_7 :: KeySym

xK_braille_dot_8 :: KeySym

xK_braille_dot_9 :: KeySym

xK_braille_dot_10 :: KeySym

xK_braille_blank :: KeySym

xK_braille_dots_1 :: KeySym

xK_braille_dots_2 :: KeySym

xK_braille_dots_12 :: KeySym

xK_braille_dots_3 :: KeySym

xK_braille_dots_13 :: KeySym

xK_braille_dots_23 :: KeySym

xK_braille_dots_123 :: KeySym

xK_braille_dots_4 :: KeySym

xK_braille_dots_14 :: KeySym

xK_braille_dots_24 :: KeySym

xK_braille_dots_124 :: KeySym

xK_braille_dots_34 :: KeySym

xK_braille_dots_134 :: KeySym

xK_braille_dots_234 :: KeySym

xK_braille_dots_1234 :: KeySym

xK_braille_dots_5 :: KeySym

xK_braille_dots_15 :: KeySym

xK_braille_dots_25 :: KeySym

xK_braille_dots_125 :: KeySym

xK_braille_dots_35 :: KeySym

xK_braille_dots_135 :: KeySym

xK_braille_dots_235 :: KeySym

xK_braille_dots_1235 :: KeySym

xK_braille_dots_45 :: KeySym

xK_braille_dots_145 :: KeySym

xK_braille_dots_245 :: KeySym

xK_braille_dots_1245 :: KeySym

xK_braille_dots_345 :: KeySym

xK_braille_dots_1345 :: KeySym

xK_braille_dots_2345 :: KeySym

xK_braille_dots_12345 :: KeySym

xK_braille_dots_6 :: KeySym

xK_braille_dots_16 :: KeySym

xK_braille_dots_26 :: KeySym

xK_braille_dots_126 :: KeySym

xK_braille_dots_36 :: KeySym

xK_braille_dots_136 :: KeySym

xK_braille_dots_236 :: KeySym

xK_braille_dots_1236 :: KeySym

xK_braille_dots_46 :: KeySym

xK_braille_dots_146 :: KeySym

xK_braille_dots_246 :: KeySym

xK_braille_dots_1246 :: KeySym

xK_braille_dots_346 :: KeySym

xK_braille_dots_1346 :: KeySym

xK_braille_dots_2346 :: KeySym

xK_braille_dots_12346 :: KeySym

xK_braille_dots_56 :: KeySym

xK_braille_dots_156 :: KeySym

xK_braille_dots_256 :: KeySym

xK_braille_dots_1256 :: KeySym

xK_braille_dots_356 :: KeySym

xK_braille_dots_1356 :: KeySym

xK_braille_dots_2356 :: KeySym

xK_braille_dots_12356 :: KeySym

xK_braille_dots_456 :: KeySym

xK_braille_dots_1456 :: KeySym

xK_braille_dots_2456 :: KeySym

xK_braille_dots_12456 :: KeySym

xK_braille_dots_3456 :: KeySym

xK_braille_dots_13456 :: KeySym

xK_braille_dots_23456 :: KeySym

xK_braille_dots_123456 :: KeySym

xK_braille_dots_7 :: KeySym

xK_braille_dots_17 :: KeySym

xK_braille_dots_27 :: KeySym

xK_braille_dots_127 :: KeySym

xK_braille_dots_37 :: KeySym

xK_braille_dots_137 :: KeySym

xK_braille_dots_237 :: KeySym

xK_braille_dots_1237 :: KeySym

xK_braille_dots_47 :: KeySym

xK_braille_dots_147 :: KeySym

xK_braille_dots_247 :: KeySym

xK_braille_dots_1247 :: KeySym

xK_braille_dots_347 :: KeySym

xK_braille_dots_1347 :: KeySym

xK_braille_dots_2347 :: KeySym

xK_braille_dots_12347 :: KeySym

xK_braille_dots_57 :: KeySym

xK_braille_dots_157 :: KeySym

xK_braille_dots_257 :: KeySym

xK_braille_dots_1257 :: KeySym

xK_braille_dots_357 :: KeySym

xK_braille_dots_1357 :: KeySym

xK_braille_dots_2357 :: KeySym

xK_braille_dots_12357 :: KeySym

xK_braille_dots_457 :: KeySym

xK_braille_dots_1457 :: KeySym

xK_braille_dots_2457 :: KeySym

xK_braille_dots_12457 :: KeySym

xK_braille_dots_3457 :: KeySym

xK_braille_dots_13457 :: KeySym

xK_braille_dots_23457 :: KeySym

xK_braille_dots_123457 :: KeySym

xK_braille_dots_67 :: KeySym

xK_braille_dots_167 :: KeySym

xK_braille_dots_267 :: KeySym

xK_braille_dots_1267 :: KeySym

xK_braille_dots_367 :: KeySym

xK_braille_dots_1367 :: KeySym

xK_braille_dots_2367 :: KeySym

xK_braille_dots_12367 :: KeySym

xK_braille_dots_467 :: KeySym

xK_braille_dots_1467 :: KeySym

xK_braille_dots_2467 :: KeySym

xK_braille_dots_12467 :: KeySym

xK_braille_dots_3467 :: KeySym

xK_braille_dots_13467 :: KeySym

xK_braille_dots_23467 :: KeySym

xK_braille_dots_123467 :: KeySym

xK_braille_dots_567 :: KeySym

xK_braille_dots_1567 :: KeySym

xK_braille_dots_2567 :: KeySym

xK_braille_dots_12567 :: KeySym

xK_braille_dots_3567 :: KeySym

xK_braille_dots_13567 :: KeySym

xK_braille_dots_23567 :: KeySym

xK_braille_dots_123567 :: KeySym

xK_braille_dots_4567 :: KeySym

xK_braille_dots_14567 :: KeySym

xK_braille_dots_24567 :: KeySym

xK_braille_dots_124567 :: KeySym

xK_braille_dots_34567 :: KeySym

xK_braille_dots_134567 :: KeySym

xK_braille_dots_234567 :: KeySym

xK_braille_dots_1234567 :: KeySym

xK_braille_dots_8 :: KeySym

xK_braille_dots_18 :: KeySym

xK_braille_dots_28 :: KeySym

xK_braille_dots_128 :: KeySym

xK_braille_dots_38 :: KeySym

xK_braille_dots_138 :: KeySym

xK_braille_dots_238 :: KeySym

xK_braille_dots_1238 :: KeySym

xK_braille_dots_48 :: KeySym

xK_braille_dots_148 :: KeySym

xK_braille_dots_248 :: KeySym

xK_braille_dots_1248 :: KeySym

xK_braille_dots_348 :: KeySym

xK_braille_dots_1348 :: KeySym

xK_braille_dots_2348 :: KeySym

xK_braille_dots_12348 :: KeySym

xK_braille_dots_58 :: KeySym

xK_braille_dots_158 :: KeySym

xK_braille_dots_258 :: KeySym

xK_braille_dots_1258 :: KeySym

xK_braille_dots_358 :: KeySym

xK_braille_dots_1358 :: KeySym

xK_braille_dots_2358 :: KeySym

xK_braille_dots_12358 :: KeySym

xK_braille_dots_458 :: KeySym

xK_braille_dots_1458 :: KeySym

xK_braille_dots_2458 :: KeySym

xK_braille_dots_12458 :: KeySym

xK_braille_dots_3458 :: KeySym

xK_braille_dots_13458 :: KeySym

xK_braille_dots_23458 :: KeySym

xK_braille_dots_123458 :: KeySym

xK_braille_dots_68 :: KeySym

xK_braille_dots_168 :: KeySym

xK_braille_dots_268 :: KeySym

xK_braille_dots_1268 :: KeySym

xK_braille_dots_368 :: KeySym

xK_braille_dots_1368 :: KeySym

xK_braille_dots_2368 :: KeySym

xK_braille_dots_12368 :: KeySym

xK_braille_dots_468 :: KeySym

xK_braille_dots_1468 :: KeySym

xK_braille_dots_2468 :: KeySym

xK_braille_dots_12468 :: KeySym

xK_braille_dots_3468 :: KeySym

xK_braille_dots_13468 :: KeySym

xK_braille_dots_23468 :: KeySym

xK_braille_dots_123468 :: KeySym

xK_braille_dots_568 :: KeySym

xK_braille_dots_1568 :: KeySym

xK_braille_dots_2568 :: KeySym

xK_braille_dots_12568 :: KeySym

xK_braille_dots_3568 :: KeySym

xK_braille_dots_13568 :: KeySym

xK_braille_dots_23568 :: KeySym

xK_braille_dots_123568 :: KeySym

xK_braille_dots_4568 :: KeySym

xK_braille_dots_14568 :: KeySym

xK_braille_dots_24568 :: KeySym

xK_braille_dots_124568 :: KeySym

xK_braille_dots_34568 :: KeySym

xK_braille_dots_134568 :: KeySym

xK_braille_dots_234568 :: KeySym

xK_braille_dots_1234568 :: KeySym

xK_braille_dots_78 :: KeySym

xK_braille_dots_178 :: KeySym

xK_braille_dots_278 :: KeySym

xK_braille_dots_1278 :: KeySym

xK_braille_dots_378 :: KeySym

xK_braille_dots_1378 :: KeySym

xK_braille_dots_2378 :: KeySym

xK_braille_dots_12378 :: KeySym

xK_braille_dots_478 :: KeySym

xK_braille_dots_1478 :: KeySym

xK_braille_dots_2478 :: KeySym

xK_braille_dots_12478 :: KeySym

xK_braille_dots_3478 :: KeySym

xK_braille_dots_13478 :: KeySym

xK_braille_dots_23478 :: KeySym

xK_braille_dots_123478 :: KeySym

xK_braille_dots_578 :: KeySym

xK_braille_dots_1578 :: KeySym

xK_braille_dots_2578 :: KeySym

xK_braille_dots_12578 :: KeySym

xK_braille_dots_3578 :: KeySym

xK_braille_dots_13578 :: KeySym

xK_braille_dots_23578 :: KeySym

xK_braille_dots_123578 :: KeySym

xK_braille_dots_4578 :: KeySym

xK_braille_dots_14578 :: KeySym

xK_braille_dots_24578 :: KeySym

xK_braille_dots_124578 :: KeySym

xK_braille_dots_34578 :: KeySym

xK_braille_dots_134578 :: KeySym

xK_braille_dots_234578 :: KeySym

xK_braille_dots_1234578 :: KeySym

xK_braille_dots_678 :: KeySym

xK_braille_dots_1678 :: KeySym

xK_braille_dots_2678 :: KeySym

xK_braille_dots_12678 :: KeySym

xK_braille_dots_3678 :: KeySym

xK_braille_dots_13678 :: KeySym

xK_braille_dots_23678 :: KeySym

xK_braille_dots_123678 :: KeySym

xK_braille_dots_4678 :: KeySym

xK_braille_dots_14678 :: KeySym

xK_braille_dots_24678 :: KeySym

xK_braille_dots_124678 :: KeySym

xK_braille_dots_34678 :: KeySym

xK_braille_dots_134678 :: KeySym

xK_braille_dots_234678 :: KeySym

xK_braille_dots_1234678 :: KeySym

xK_braille_dots_5678 :: KeySym

xK_braille_dots_15678 :: KeySym

xK_braille_dots_25678 :: KeySym

xK_braille_dots_125678 :: KeySym

xK_braille_dots_35678 :: KeySym

xK_braille_dots_135678 :: KeySym

xK_braille_dots_235678 :: KeySym

xK_braille_dots_1235678 :: KeySym

xK_braille_dots_45678 :: KeySym

xK_braille_dots_145678 :: KeySym

xK_braille_dots_245678 :: KeySym

xK_braille_dots_1245678 :: KeySym

xK_braille_dots_345678 :: KeySym

xK_braille_dots_1345678 :: KeySym

xK_braille_dots_2345678 :: KeySym

xK_braille_dots_12345678 :: KeySym