Graphics.X11.Types

Description

A collection of type declarations for interfacing with X11.

Documentation

type XID = Word64

type Mask = Word64

type Atom = Word64

type VisualID = Word64

type Time = Word64

type Window = XID

type Drawable = XID

type Font = XID

type Pixmap = XID

type Cursor = XID

type Colormap = XID

type GContext = XID

type KeyCode = Word8

type SizeID = Word16

type SubpixelOrder = Word16

type Connection = Word16

type RROutput = Word64

type RRCrtc = Word64

type RRMode = Word64

type XRRModeFlags = Word64

Enumeration types

These types were introduced to make function types clearer. Note that the types are synonyms for Int, so no extra typesafety was obtained.

Key symbols

type KeySym = XID

xK_VoidSymbol :: KeySym

xK_BackSpace :: KeySym

xK_Tab :: KeySym

xK_Linefeed :: KeySym

xK_Clear :: KeySym

xK_Return :: KeySym

xK_Pause :: KeySym

xK_Scroll_Lock :: KeySym

xK_Sys_Req :: KeySym

xK_Escape :: KeySym

xK_Delete :: KeySym

xK_Multi_key :: KeySym

xK_Codeinput :: KeySym

xK_SingleCandidate :: KeySym

xK_MultipleCandidate :: KeySym

xK_PreviousCandidate :: KeySym

xK_Home :: KeySym

xK_Left :: KeySym

xK_Up :: KeySym

xK_Right :: KeySym

xK_Down :: KeySym

xK_Prior :: KeySym

xK_Page_Up :: KeySym

xK_Next :: KeySym

xK_Page_Down :: KeySym

xK_End :: KeySym

xK_Begin :: KeySym

xK_Select :: KeySym

xK_Print :: KeySym

xK_Execute :: KeySym

xK_Insert :: KeySym

xK_Undo :: KeySym

xK_Redo :: KeySym

xK_Menu :: KeySym

xK_Find :: KeySym

xK_Cancel :: KeySym

xK_Help :: KeySym

xK_Break :: KeySym

xK_Mode_switch :: KeySym

xK_script_switch :: KeySym

xK_Num_Lock :: KeySym

xK_KP_Space :: KeySym

xK_KP_Tab :: KeySym

xK_KP_Enter :: KeySym

xK_KP_F1 :: KeySym

xK_KP_F2 :: KeySym

xK_KP_F3 :: KeySym

xK_KP_F4 :: KeySym

xK_KP_Home :: KeySym

xK_KP_Left :: KeySym

xK_KP_Up :: KeySym

xK_KP_Right :: KeySym

xK_KP_Down :: KeySym

xK_KP_Prior :: KeySym

xK_KP_Page_Up :: KeySym

xK_KP_Next :: KeySym

xK_KP_Page_Down :: KeySym

xK_KP_End :: KeySym

xK_KP_Begin :: KeySym

xK_KP_Insert :: KeySym

xK_KP_Delete :: KeySym

xK_KP_Equal :: KeySym

xK_KP_Multiply :: KeySym

xK_KP_Add :: KeySym

xK_KP_Separator :: KeySym

xK_KP_Subtract :: KeySym

xK_KP_Decimal :: KeySym

xK_KP_Divide :: KeySym

xK_KP_0 :: KeySym

xK_KP_1 :: KeySym

xK_KP_2 :: KeySym

xK_KP_3 :: KeySym

xK_KP_4 :: KeySym

xK_KP_5 :: KeySym

xK_KP_6 :: KeySym

xK_KP_7 :: KeySym

xK_KP_8 :: KeySym

xK_KP_9 :: KeySym

xK_F1 :: KeySym

xK_F2 :: KeySym

xK_F3 :: KeySym

xK_F4 :: KeySym

xK_F5 :: KeySym

xK_F6 :: KeySym

xK_F7 :: KeySym

xK_F8 :: KeySym

xK_F9 :: KeySym

xK_F10 :: KeySym

xK_F11 :: KeySym

xK_L1 :: KeySym

xK_F12 :: KeySym

xK_L2 :: KeySym

xK_F13 :: KeySym

xK_L3 :: KeySym

xK_F14 :: KeySym

xK_L4 :: KeySym

xK_F15 :: KeySym

xK_L5 :: KeySym

xK_F16 :: KeySym

xK_L6 :: KeySym

xK_F17 :: KeySym

xK_L7 :: KeySym

xK_F18 :: KeySym

xK_L8 :: KeySym

xK_F19 :: KeySym

xK_L9 :: KeySym

xK_F20 :: KeySym

xK_L10 :: KeySym

xK_F21 :: KeySym

xK_R1 :: KeySym

xK_F22 :: KeySym

xK_R2 :: KeySym

xK_F23 :: KeySym

xK_R3 :: KeySym

xK_F24 :: KeySym

xK_R4 :: KeySym

xK_F25 :: KeySym

xK_R5 :: KeySym

xK_F26 :: KeySym

xK_R6 :: KeySym

xK_F27 :: KeySym

xK_R7 :: KeySym

xK_F28 :: KeySym

xK_R8 :: KeySym

xK_F29 :: KeySym

xK_R9 :: KeySym

xK_F30 :: KeySym

xK_R10 :: KeySym

xK_F31 :: KeySym

xK_R11 :: KeySym

xK_F32 :: KeySym

xK_R12 :: KeySym

xK_F33 :: KeySym

xK_R13 :: KeySym

xK_F34 :: KeySym

xK_R14 :: KeySym

xK_F35 :: KeySym

xK_R15 :: KeySym

xK_Shift_L :: KeySym

xK_Shift_R :: KeySym

xK_Control_L :: KeySym

xK_Control_R :: KeySym

xK_Caps_Lock :: KeySym

xK_Shift_Lock :: KeySym

xK_Meta_L :: KeySym

xK_Meta_R :: KeySym

xK_Alt_L :: KeySym

xK_Alt_R :: KeySym

xK_Super_L :: KeySym

xK_Super_R :: KeySym

xK_Hyper_L :: KeySym

xK_Hyper_R :: KeySym

xK_space :: KeySym

xK_exclam :: KeySym

xK_quotedbl :: KeySym

xK_numbersign :: KeySym

xK_dollar :: KeySym

xK_percent :: KeySym

xK_ampersand :: KeySym

xK_apostrophe :: KeySym

xK_quoteright :: KeySym

xK_parenleft :: KeySym

xK_parenright :: KeySym

xK_asterisk :: KeySym

xK_plus :: KeySym

xK_comma :: KeySym

xK_minus :: KeySym

xK_period :: KeySym

xK_slash :: KeySym

xK_0 :: KeySym

xK_1 :: KeySym

xK_2 :: KeySym

xK_3 :: KeySym

xK_4 :: KeySym

xK_5 :: KeySym

xK_6 :: KeySym

xK_7 :: KeySym

xK_8 :: KeySym

xK_9 :: KeySym

xK_colon :: KeySym

xK_semicolon :: KeySym

xK_less :: KeySym

xK_equal :: KeySym

xK_greater :: KeySym

xK_question :: KeySym

xK_at :: KeySym

xK_A :: KeySym

xK_B :: KeySym

xK_C :: KeySym

xK_D :: KeySym

xK_E :: KeySym

xK_F :: KeySym

xK_G :: KeySym

xK_H :: KeySym

xK_I :: KeySym

xK_J :: KeySym

xK_K :: KeySym

xK_L :: KeySym

xK_M :: KeySym

xK_N :: KeySym

xK_O :: KeySym

xK_P :: KeySym

xK_Q :: KeySym

xK_R :: KeySym

xK_S :: KeySym

xK_T :: KeySym

xK_U :: KeySym

xK_V :: KeySym

xK_W :: KeySym

xK_X :: KeySym

xK_Y :: KeySym

xK_Z :: KeySym

xK_bracketleft :: KeySym

xK_backslash :: KeySym

xK_bracketright :: KeySym

xK_asciicircum :: KeySym

xK_underscore :: KeySym

xK_grave :: KeySym

xK_quoteleft :: KeySym

xK_a :: KeySym

xK_b :: KeySym

xK_c :: KeySym

xK_d :: KeySym

xK_e :: KeySym

xK_f :: KeySym

xK_g :: KeySym

xK_h :: KeySym

xK_i :: KeySym

xK_j :: KeySym

xK_k :: KeySym

xK_l :: KeySym

xK_m :: KeySym

xK_n :: KeySym

xK_o :: KeySym

xK_p :: KeySym

xK_q :: KeySym

xK_r :: KeySym

xK_s :: KeySym

xK_t :: KeySym

xK_u :: KeySym

xK_v :: KeySym

xK_w :: KeySym

xK_x :: KeySym

xK_y :: KeySym

xK_z :: KeySym

xK_braceleft :: KeySym

xK_bar :: KeySym

xK_braceright :: KeySym

xK_asciitilde :: KeySym

xK_nobreakspace :: KeySym

xK_exclamdown :: KeySym

xK_cent :: KeySym

xK_sterling :: KeySym

xK_currency :: KeySym

xK_yen :: KeySym

xK_brokenbar :: KeySym

xK_section :: KeySym

xK_diaeresis :: KeySym

xK_copyright :: KeySym

xK_ordfeminine :: KeySym

xK_guillemotleft :: KeySym

xK_notsign :: KeySym

xK_hyphen :: KeySym

xK_registered :: KeySym

xK_macron :: KeySym

xK_degree :: KeySym

xK_plusminus :: KeySym

xK_twosuperior :: KeySym

xK_threesuperior :: KeySym

xK_acute :: KeySym

xK_mu :: KeySym

xK_paragraph :: KeySym

xK_periodcentered :: KeySym

xK_cedilla :: KeySym

xK_onesuperior :: KeySym

xK_masculine :: KeySym

xK_guillemotright :: KeySym

xK_onequarter :: KeySym

xK_onehalf :: KeySym

xK_threequarters :: KeySym

xK_questiondown :: KeySym

xK_Agrave :: KeySym

xK_Aacute :: KeySym

xK_Acircumflex :: KeySym

xK_Atilde :: KeySym

xK_Adiaeresis :: KeySym

xK_Aring :: KeySym

xK_AE :: KeySym

xK_Ccedilla :: KeySym

xK_Egrave :: KeySym

xK_Eacute :: KeySym

xK_Ecircumflex :: KeySym

xK_Ediaeresis :: KeySym

xK_Igrave :: KeySym

xK_Iacute :: KeySym

xK_Icircumflex :: KeySym

xK_Idiaeresis :: KeySym

xK_ETH :: KeySym

xK_Eth :: KeySym

xK_Ntilde :: KeySym

xK_Ograve :: KeySym

xK_Oacute :: KeySym

xK_Ocircumflex :: KeySym

xK_Otilde :: KeySym

xK_Odiaeresis :: KeySym

xK_multiply :: KeySym

xK_Ooblique :: KeySym

xK_Ugrave :: KeySym

xK_Uacute :: KeySym

xK_Ucircumflex :: KeySym

xK_Udiaeresis :: KeySym

xK_Yacute :: KeySym

xK_THORN :: KeySym

xK_Thorn :: KeySym

xK_ssharp :: KeySym

xK_agrave :: KeySym

xK_aacute :: KeySym

xK_acircumflex :: KeySym

xK_atilde :: KeySym

xK_adiaeresis :: KeySym

xK_aring :: KeySym

xK_ae :: KeySym

xK_ccedilla :: KeySym

xK_egrave :: KeySym

xK_eacute :: KeySym

xK_ecircumflex :: KeySym

xK_ediaeresis :: KeySym

xK_igrave :: KeySym

xK_iacute :: KeySym

xK_icircumflex :: KeySym

xK_idiaeresis :: KeySym

xK_eth :: KeySym

xK_ntilde :: KeySym

xK_ograve :: KeySym

xK_oacute :: KeySym

xK_ocircumflex :: KeySym

xK_otilde :: KeySym

xK_odiaeresis :: KeySym

xK_division :: KeySym

xK_oslash :: KeySym

xK_ugrave :: KeySym

xK_uacute :: KeySym

xK_ucircumflex :: KeySym

xK_udiaeresis :: KeySym

xK_yacute :: KeySym

xK_thorn :: KeySym

xK_ydiaeresis :: KeySym

Event masks

type EventMask = Mask

noEventMask :: EventMask

keyPressMask :: EventMask

keyReleaseMask :: EventMask

buttonPressMask :: EventMask

buttonReleaseMask :: EventMask

enterWindowMask :: EventMask

leaveWindowMask :: EventMask

pointerMotionMask :: EventMask

pointerMotionHintMask :: EventMask

button1MotionMask :: EventMask

button2MotionMask :: EventMask

button3MotionMask :: EventMask

button4MotionMask :: EventMask

button5MotionMask :: EventMask

buttonMotionMask :: EventMask

keymapStateMask :: EventMask

exposureMask :: EventMask

visibilityChangeMask :: EventMask

structureNotifyMask :: EventMask

resizeRedirectMask :: EventMask

substructureNotifyMask :: EventMask

substructureRedirectMask :: EventMask

focusChangeMask :: EventMask

propertyChangeMask :: EventMask

colormapChangeMask :: EventMask

ownerGrabButtonMask :: EventMask

rrScreenChangeNotifyMask :: EventMask

rrCrtcChangeNotifyMask :: EventMask

rrOutputChangeNotifyMask :: EventMask

rrOutputPropertyNotifyMask :: EventMask

Event types

type EventType = Word32

keyPress :: EventType

keyRelease :: EventType

buttonPress :: EventType

buttonRelease :: EventType

motionNotify :: EventType

enterNotify :: EventType

leaveNotify :: EventType

focusIn :: EventType

focusOut :: EventType

keymapNotify :: EventType

expose :: EventType

graphicsExpose :: EventType

noExpose :: EventType

visibilityNotify :: EventType

createNotify :: EventType

destroyNotify :: EventType

unmapNotify :: EventType

mapNotify :: EventType

mapRequest :: EventType

reparentNotify :: EventType

configureNotify :: EventType

configureRequest :: EventType

gravityNotify :: EventType

resizeRequest :: EventType

circulateNotify :: EventType

circulateRequest :: EventType

propertyNotify :: EventType

selectionClear :: EventType

selectionRequest :: EventType

selectionNotify :: EventType

colormapNotify :: EventType

clientMessage :: EventType

mappingNotify :: EventType

rrScreenChangeNotify :: EventType

rrNotify :: EventType

rrNotifyCrtcChange :: EventType

rrNotifyOutputChange :: EventType

rrNotifyOutputProperty :: EventType

lASTEvent :: EventType

Modifiers

type Modifier = CUInt

shiftMapIndex :: Modifier

lockMapIndex :: Modifier

controlMapIndex :: Modifier

mod1MapIndex :: Modifier

mod2MapIndex :: Modifier

mod3MapIndex :: Modifier

mod4MapIndex :: Modifier

mod5MapIndex :: Modifier

anyModifier :: Modifier

Key masks

type KeyMask = Modifier

noModMask :: KeyMask

shiftMask :: KeyMask

lockMask :: KeyMask

controlMask :: KeyMask

mod1Mask :: KeyMask

mod2Mask :: KeyMask

mod3Mask :: KeyMask

mod4Mask :: KeyMask

mod5Mask :: KeyMask

Button masks

type ButtonMask = Modifier

button1Mask :: ButtonMask

button2Mask :: ButtonMask

button3Mask :: ButtonMask

button4Mask :: ButtonMask

button5Mask :: ButtonMask

Buttons

type Button = Word32

button1 :: Button

button2 :: Button

button3 :: Button

button4 :: Button

button5 :: Button

Notify modes

type NotifyMode = CInt

notifyNormal :: NotifyMode

notifyGrab :: NotifyMode

notifyUngrab :: NotifyMode

notifyWhileGrabbed :: NotifyMode

notifyHint :: NotifyMode

Notify details

type NotifyDetail = CInt

notifyAncestor :: NotifyDetail

notifyVirtual :: NotifyDetail

notifyInferior :: NotifyDetail

notifyNonlinear :: NotifyDetail

notifyNonlinearVirtual :: NotifyDetail

notifyPointer :: NotifyDetail

notifyPointerRoot :: NotifyDetail

notifyDetailNone :: NotifyDetail

Visibility

type Visibility = CInt

visibilityUnobscured :: Visibility

visibilityPartiallyObscured :: Visibility

visibilityFullyObscured :: Visibility

Place of window

type Place = CInt

Place of window relative to siblings (used in Circulation requests or events)

placeOnTop :: Place

placeOnBottom :: Place

Protocols

type Protocol = CInt

familyInternet :: Protocol

familyDECnet :: Protocol

familyChaos :: Protocol

Property notification

type PropertyNotification = CInt

propertyNewValue :: PropertyNotification

propertyDelete :: PropertyNotification

Colormap notification

type ColormapNotification = CInt

colormapUninstalled :: ColormapNotification

colormapInstalled :: ColormapNotification

Grab modes

type GrabMode = CInt

grabModeSync :: GrabMode

grabModeAsync :: GrabMode

Grab status

type GrabStatus = CInt

grabSuccess :: GrabStatus

alreadyGrabbed :: GrabStatus

grabInvalidTime :: GrabStatus

grabNotViewable :: GrabStatus

grabFrozen :: GrabStatus

Allow events

type AllowEvents = CInt

asyncPointer :: AllowEvents

syncPointer :: AllowEvents

replayPointer :: AllowEvents

asyncKeyboard :: AllowEvents

syncKeyboard :: AllowEvents

replayKeyboard :: AllowEvents

asyncBoth :: AllowEvents

syncBoth :: AllowEvents

Focus modes

type FocusMode = CInt

revertToNone :: FocusMode

revertToPointerRoot :: FocusMode

revertToParent :: FocusMode

Error codes

type ErrorCode = CInt

success :: ErrorCode

badRequest :: ErrorCode

badValue :: ErrorCode

badWindow :: ErrorCode

badPixmap :: ErrorCode

badAtom :: ErrorCode

badCursor :: ErrorCode

badFont :: ErrorCode

badMatch :: ErrorCode

badDrawable :: ErrorCode

badAccess :: ErrorCode

badAlloc :: ErrorCode

badColor :: ErrorCode

badGC :: ErrorCode

Xlib functions with return values of type Status return zero on failure and nonzero on success.

badIDChoice :: ErrorCode

badName :: ErrorCode

badLength :: ErrorCode

badImplementation :: ErrorCode

firstExtensionError :: ErrorCode

lastExtensionError :: ErrorCode

Return status

type Status = CInt

throwIfZero :: String -> IO Status -> IO ()

WindowClass

type WindowClass = CInt

copyFromParent :: WindowClass

inputOutput :: WindowClass

inputOnly :: WindowClass

Attribute masks

type AttributeMask = Mask

cWBackPixmap :: AttributeMask

cWBackPixel :: AttributeMask

cWBorderPixmap :: AttributeMask

cWBorderPixel :: AttributeMask

cWBitGravity :: AttributeMask

cWWinGravity :: AttributeMask

cWBackingStore :: AttributeMask

cWBackingPlanes :: AttributeMask

cWBackingPixel :: AttributeMask

cWOverrideRedirect :: AttributeMask

cWSaveUnder :: AttributeMask

cWEventMask :: AttributeMask

cWDontPropagate :: AttributeMask

cWColormap :: AttributeMask

cWCursor :: AttributeMask

Close down modes

type CloseDownMode = CInt

destroyAll :: CloseDownMode

retainPermanent :: CloseDownMode

retainTemporary :: CloseDownMode

QueryBestSize classes

type QueryBestSizeClass = CInt

cursorShape :: QueryBestSizeClass

tileShape :: QueryBestSizeClass

stippleShape :: QueryBestSizeClass

Graphics functions

type GXFunction = CInt

gXclear :: GXFunction

gXand :: GXFunction

gXandReverse :: GXFunction

gXcopy :: GXFunction

gXandInverted :: GXFunction

gXnoop :: GXFunction

gXxor :: GXFunction

gXor :: GXFunction

gXnor :: GXFunction

gXequiv :: GXFunction

gXinvert :: GXFunction

gXorReverse :: GXFunction

gXcopyInverted :: GXFunction

gXorInverted :: GXFunction

gXnand :: GXFunction

gXset :: GXFunction

Line styles

type LineStyle = CInt

lineSolid :: LineStyle

lineOnOffDash :: LineStyle

lineDoubleDash :: LineStyle

Cap styles

type CapStyle = CInt

capNotLast :: CapStyle

capButt :: CapStyle

capRound :: CapStyle

capProjecting :: CapStyle

Join styles

type JoinStyle = CInt

joinMiter :: JoinStyle

joinRound :: JoinStyle

joinBevel :: JoinStyle

Fill styles

type FillStyle = CInt

fillSolid :: FillStyle

fillTiled :: FillStyle

fillStippled :: FillStyle

fillOpaqueStippled :: FillStyle

Fill rules

type FillRule = CInt

evenOddRule :: FillRule

windingRule :: FillRule

Subwindow modes

type SubWindowMode = CInt

clipByChildren :: SubWindowMode

includeInferiors :: SubWindowMode

Coordinate modes

type CoordinateMode = CInt

coordModeOrigin :: CoordinateMode

coordModePrevious :: CoordinateMode

Polygon shapes

type PolygonShape = CInt

complex :: PolygonShape

nonconvex :: PolygonShape

convex :: PolygonShape

Arc modes

type ArcMode = CInt

arcChord :: ArcMode

arcPieSlice :: ArcMode

GC masks

type GCMask = CInt

gCFunction :: GCMask

gCPlaneMask :: GCMask

gCForeground :: GCMask

gCBackground :: GCMask

gCLineWidth :: GCMask

gCLineStyle :: GCMask

gCCapStyle :: GCMask

gCJoinStyle :: GCMask

gCFillStyle :: GCMask

gCFillRule :: GCMask

gCTile :: GCMask

gCStipple :: GCMask

gCTileStipXOrigin :: GCMask

gCTileStipYOrigin :: GCMask

gCFont :: GCMask

gCSubwindowMode :: GCMask

gCGraphicsExposures :: GCMask

gCClipXOrigin :: GCMask

gCClipYOrigin :: GCMask

gCClipMask :: GCMask

gCDashOffset :: GCMask

gCDashList :: GCMask

gCArcMode :: GCMask

gCLastBit :: GCMask

Circulation direction

type CirculationDirection = CInt

raiseLowest :: CirculationDirection

lowerHighest :: CirculationDirection

Byte order

type ByteOrder = CInt

lSBFirst :: ByteOrder

mSBFirst :: ByteOrder

ColormapAlloc

type ColormapAlloc = CInt

allocNone :: ColormapAlloc

allocAll :: ColormapAlloc

Mapping requests

type MappingRequest = CInt

mappingModifier :: MappingRequest

mappingKeyboard :: MappingRequest

mappingPointer :: MappingRequest

ChangeSaveSetMode

type ChangeSaveSetMode = CInt

setModeInsert :: ChangeSaveSetMode

setModeDelete :: ChangeSaveSetMode

Bit gravity

type BitGravity = CInt

forgetGravity :: BitGravity

northWestGravity :: BitGravity

northGravity :: BitGravity

northEastGravity :: BitGravity

westGravity :: BitGravity

centerGravity :: BitGravity

eastGravity :: BitGravity

southWestGravity :: BitGravity

southGravity :: BitGravity

southEastGravity :: BitGravity

staticGravity :: BitGravity

Window gravity

type WindowGravity = CInt

unmapGravity :: WindowGravity

Backing store

type BackingStore = CInt

notUseful :: BackingStore

whenMapped :: BackingStore

always :: BackingStore

doRed :: Word8

doGreen :: Word8

doBlue :: Word8

Font direction

type FontDirection = CInt

fontLeftToRight :: FontDirection

fontRightToLeft :: FontDirection

Image format

type ImageFormat = CInt

xyBitmap :: ImageFormat

xyPixmap :: ImageFormat

zPixmap :: ImageFormat

Reflection and Rotation

type Rotation = Word16

type Reflection = Word16

xRR_Rotate_0 :: Rotation

xRR_Rotate_90 :: Rotation

xRR_Rotate_180 :: Rotation

xRR_Rotate_270 :: Rotation

xRR_Reflect_X :: Reflection

xRR_Reflect_Y :: Reflection

xRR_Connected :: Connection

xRR_Disconnected :: Connection

xRR_UnknownConnection :: Connection